pyproject.toml
[project]
name = "ai-lqa-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "arslantoor",email = "muhammadarslantoor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "amqp (==5.2.0)",
    "asgiref (==3.8.1)",
    "async-timeout (==4.0.3)",
    "attrs (==24.3.0)",
    "billiard (==4.2.0)",
    "celery (==5.3.6)",
    "certifi (==2024.2.2)",
    "cffi (==1.16.0)",
    "charset-normalizer (==3.3.2)",
    "click (==8.1.7)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1)",
    "click-repl (==0.3.0)",
    "cryptography (==42.0.5)",
    "defusedxml (==0.7.1)",
    "dj-rest-auth (==4.0.1)",
    "django (==5.0.4)",
    "django-allauth (==0.54.0)",
    "django-cors-headers (==4.3.1)",
    "django-debug-toolbar (==4.3.0)",
    "django-environ (==0.11.2)",
    "djangorestframework (==3.15.1)",
    "djangorestframework-simplejwt (==5.3.1)",
    "drf-spectacular (==0.28.0)",
    "drf-yasg (==1.21.8)",
    "flake8 (==7.0.0)",
    "flake8-isort (==6.1.1)",
    "idna (==3.7)",
    "inflection (==0.5.1)",
    "isort (==5.13.2)",
    "jsonschema (==4.23.0)",
    "jsonschema-specifications (==2024.10.1)",
    "kombu (==5.3.7)",
    "mccabe (==0.7.0)",
    "oauthlib (==3.2.2)",
    "packaging (==24.2)",
    "prompt-toolkit (==3.0.43)",
    "psycopg2-binary (==2.9.9)",
    "pycodestyle (==2.11.1)",
    "pycparser (==2.22)",
    "pydocstyle (==6.3.0)",
    "pyflakes (==3.2.0)",
    "pyjwt (==2.8.0)",
    "python-dateutil (==2.9.0.post0)",
    "python3-openid (==3.2.0)",
    "pytz (==2024.1)",
    "pyyaml (==6.0.2)",
    "redis (==5.0.3)",
    "referencing (==0.35.1)",
    "requests (==2.31.0)",
    "requests-oauthlib (==2.0.0)",
    "rpds-py (==0.22.3)",
    "sentry-sdk (==2.0.1)",
    "six (==1.16.0)",
    "snowballstemmer (==2.2.0)",
    "sqlparse (==0.5.0)",
    "stripe (==9.1.0)",
    "typing-extensions (==4.11.0)",
    "tzdata (==2024.1)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.2.1)",
    "vine (==5.1.0)",
    "wcwidth (==0.2.13)",
    "openai (>=1.66.3,<2.0.0)",
    "pgvector (>=0.3.6,<0.4.0)",
    "pinecone-client (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
